==> jupyter_notebooks/01 - DataCollection.ipynb <==
# **Data Collection Notebook**

## Objectives

* Fetch data from Kaggle and save it as raw data.
* Inspect the data and save it under outputs/datasets/collection

## Inputs

*   Kaggle JSON file - the authentication token.

## Outputs

* Generate Dataset: 
- outputs/datasets/collection/HousePricing.csv
- "outputs/datasets/collection/InheritedHouses.csv"


---

# Install python packages in the notebooks

%pip install -r /workspace/HeritageHousePrediction/requirements.txt

# Change working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory.
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Fetch data from Kaggle

Install Kaggle package to fetch data

%pip install kaggle==1.5.12

os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

Define the Kaggle dataset, and destination folder and download it.

KaggleDatasetPath = "codeinstitute/housing-prices-data"
DestinationFolder = "inputs/datasets/raw"   
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

Unzip the downloaded file, delete the zip file and delete the kaggle.json file

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

---

# Load and Inspect Data

import pandas as pd
df = pd.read_csv("inputs/datasets/raw/house-price-20211124T154130Z-001/house-price/house_prices_records.csv")
inherited_houses = pd.read_csv("inputs/datasets/raw/house-price-20211124T154130Z-001/house-price/inherited_houses.csv")
df.head()

inherited_houses

DataFrame Summary

df.info()

inherited_houses.info()

# Push files to Repo

import os
try:
  # create here your folder
  os.makedirs(name='outputs/datasets/collection')
except Exception as e:
  print(e)

df.to_csv('outputs/datasets/collection/HousePricing.csv', index=False)

inherited_houses.to_csv('outputs/datasets/collection/InheritedHouses.csv',index=False)
