==> jupyter_notebooks/03 - DataCleaning.ipynb <==
# Data Cleaning Notebook

## Objectives

*   Evaluate missing data
*   Clean data

## Inputs

* outputs/datasets/collection/HousePricing.csv


## Conclusions

  Data Cleaning Pipeline

  * Median Imputer - `['2ndFlrSF', 'BedroomAbvGr', 'GarageYrBlt', 'LotFrontage', 'MasVnrArea']` 

  * Categorical Imputer - `['BsmtFinType1','GarageFinish']`. 

  * Drop - `['EnclosedPorch', 'WoodDeckSF']` 



---


## Change working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import ppscore as pps
import warnings
%matplotlib inline

warnings.filterwarnings('ignore')

from sklearn.pipeline import Pipeline
from ydata_profiling import ProfileReport
from feature_engine.imputation import CategoricalImputer
from feature_engine.imputation import MeanMedianImputer
from feature_engine.selection import DropFeatures


# Load Collected data

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

df = pd.read_csv("outputs/datasets/collection/HousePricing.csv")
df.head(20)

# Data Exploration

In Data Cleaning you are interested to check the distribution and shape of a variable with missing data.

vars_with_missing_data = df.columns[df.isna().sum() > 0].to_list()
vars_with_missing_data

from ydata_profiling import ProfileReport
if vars_with_missing_data:
    profile = ProfileReport(df=df[vars_with_missing_data], minimal=True)
    profile.to_notebook_iframe()
else:
    print("There are no variables with missing data")


# Data Cleaning

## Assessing Missing Data Levels

* Custom function to display missing data levels in a DataFrame, it shows the absolute levels, relative levels and data type.

def EvaluateMissingData(df):
    missing_data_absolute = df.isnull().sum()
    missing_data_percentage = round(missing_data_absolute/len(df)*100, 2)
    df_missing_data = (pd.DataFrame(
                            data={"RowsWithMissingData": missing_data_absolute,
                                   "PercentageOfDataset": missing_data_percentage,
                                   "DataType": df.dtypes}
                                    )
                          .sort_values(by=['PercentageOfDataset'], ascending=False)
                          .query("PercentageOfDataset > 0")
                          )

    return df_missing_data


Check missing data levels for the collected dataset.

EvaluateMissingData(df)

## Dealing with Missing Data

* This is a custom function from the 'Feature Engineering Lesson Notebook unit 9'

def DataCleaningEffect(df_original,df_cleaned,variables_applied_with_method):

  flag_count=1 # Indicate plot number
  
  # distinguish between numerical and categorical variables
  categorical_variables = df_original.select_dtypes(exclude=['number']).columns 

  # scan over variables, 
    # first on variables that you applied the method
    # if the variable is a numerical plot, a histogram if categorical plot a barplot
  for set_of_variables in [variables_applied_with_method]:
    print("\n=====================================================================================")
    print(f"* Distribution Effect Analysis After Data Cleaning Method in the following variables:")
    print(f"{set_of_variables} \n\n")
  

    for var in set_of_variables:
      if var in categorical_variables:  # it is categorical variable: barplot
        
        df1 = pd.DataFrame({"Type":"Original","Value":df_original[var]})
        df2 = pd.DataFrame({"Type":"Cleaned","Value":df_cleaned[var]})
        dfAux = pd.concat([df1, df2], axis=0)
        fig , axes = plt.subplots(figsize=(15, 5))
        sns.countplot(hue='Type', data=dfAux, x="Value",palette=['#432371',"#FAAE7B"])
        axes.set(title=f"Distribution Plot {flag_count}: {var}")
        plt.xticks(rotation=90)
        plt.legend() 

      else: # it is numerical variable: histogram

        fig , axes = plt.subplots(figsize=(10, 5))
        sns.histplot(data=df_original, x=var, color="#432371", label='Original', kde=True,element="step", ax=axes)
        sns.histplot(data=df_cleaned, x=var, color="#FAAE7B", label='Cleaned', kde=True,element="step", ax=axes)
        axes.set(title=f"Distribution Plot {flag_count}: {var}")
        plt.legend() 

      plt.show()
      flag_count+= 1

### Data Cleaning Summary

Potential Cleaning approaches 

* Median Imputer - `['2ndFlrSF', 'BedroomAbvGr', 'GarageYrBlt', 'LotFrontage', 'MasVnrArea']` since the distribution isn't bell shape curved

* Categorical Imputer - `['BsmtFinType1','GarageFinish']`. 

* Drop - `['EnclosedPorch', 'WoodDeckSF']` Have a high percentage of missing values, in real life not all houses do have a porch nor a wood deck area but since we don't actually know if they don't have a porch or wood deck or that there just isn't any record of it. There for the decision to drop. 



### Split Train and Test Set

from sklearn.model_selection import train_test_split
TrainSet, TestSet, _, __ = train_test_split(
                                        df,
                                        df['SalePrice'],
                                        test_size=0.2,
                                        random_state=0)

print(f"TrainSet shape: {TrainSet.shape} \nTestSet shape: {TestSet.shape}")

df_missing_data = EvaluateMissingData(TrainSet)
print(f"* There are {df_missing_data.shape[0]} variables with missing data \n")
df_missing_data

median_imp =  ['2ndFlrSF', 'BedroomAbvGr', 'GarageYrBlt', 'LotFrontage', 'MasVnrArea']
drop = ['EnclosedPorch', 'WoodDeckSF',]
missing_cat = ['BsmtFinType1','GarageFinish']

---

### MeanMedianImputer

**Impute missing values with median** 

Since the missing data levels isn't so high and the distribution isn't bell shape curved I'll fill the missing values with median imputation
- `['2ndFlrSF', 'BedroomAbvGr', 'GarageYrBlt', 'LotFrontage', 'MasVnrArea']`

variables_method = median_imp
imputer = MeanMedianImputer(imputation_method='median', variables=variables_method)
df_method = imputer.fit_transform(TrainSet)

DataCleaningEffect(df_original=df,
                   df_cleaned=df_method,
                   variables_applied_with_method=variables_method)

The imputation hasn't had a major affect on the distribution just a little bit as expected.

### Categorical Imputation

**Imputation method 'missing'**
- `['BsmtFinType1','GarageFinish']`

variables_method = missing_cat
imputer = CategoricalImputer(imputation_method='missing', fill_value='No Record', variables=variables_method)
df_method =imputer.fit_transform(TrainSet)

DataCleaningEffect(df_original=df,
                   df_cleaned=df_method,
                   variables_applied_with_method=variables_method)

The distribution has been impacted as thought but for now i will apply this changes to prepare the model for data similar to the real-time data, 
from my research real estate house data is usually right skewed towards the higher prices

* Step 5: If you are satisfied, apply the transformation to your data.

pipeline = Pipeline([
      ('Median', MeanMedianImputer(imputation_method='median', 
                                   variables=['2ndFlrSF', 'BedroomAbvGr', 'GarageYrBlt', 
                                              'LotFrontage', 'MasVnrArea']
                                              )),
      ('CategoricalImputer', CategoricalImputer(imputation_method='missing', 
                                                fill_value='No Record', 
                                                variables=['BsmtFinType1','GarageFinish']
                                                )),
      ('Drop', DropFeatures(features_to_drop=['EnclosedPorch', 'WoodDeckSF']))
])
pipeline

### Data Cleaning Pipeline

### Fit pipeline

pipeline.fit(TrainSet)

TrainSet, TestSet = pipeline.transform(TrainSet) , pipeline.transform(TestSet)

I'll Also apply this step to the entire dataset and save for use in the streamlit dashboard. So client can inspect the dataset as used.

pipeline.fit(df)

df = pipeline.transform(df)

**Evaluation**

EvaluateMissingData(TrainSet)

EvaluateMissingData(TestSet)

EvaluateMissingData(df)

This steps will be added to the ML pipeline

### Push cleaned data to Repo

try:
  # create here your folder
  os.makedirs(name='outputs/datasets/collection/cleaned')
except Exception as e:
  print(e)


### Dataset

df.to_csv("outputs/datasets/collection/cleaned/CleanedHousePricing.csv", index=False)

### Train Set

TrainSet.to_csv("outputs/datasets/collection/cleaned/TrainSet.csv", index=False)

### Test Set 

TestSet.to_csv("outputs/datasets/collection/cleaned/TestSet.csv", index=False)
